Move Insert Bulk Copy In to own file
